feature (refs T31885): moved migration to set the static faq contents to diplanbau - as not all projects have these faqs. removed type property from PlatformFaqCategory and its migration. Adjustment of the interfaces - as some methods and constants are not shared between the entities implementing it.
