+++ demosplan/DemosPlanCoreBundle/DoctrineMigrations/2023/05/Version20230531123431.php
@@ -22,7 +22,7 @@ class Version20230531123431 extends AbstractMigration
 
         $this->addSql('CREATE TABLE platform_faq (id CHAR(36) NOT NULL, platform_faq_category_id CHAR(36) NOT NULL, title VARCHAR(255) DEFAULT \'\' NOT NULL, text LONGTEXT DEFAULT NULL, enabled TINYINT(1) DEFAULT 0 NOT NULL, create_date DATETIME NOT NULL, modify_date DATETIME NOT NULL, INDEX IDX_5ED987DF259D9BFA (platform_faq_category_id), PRIMARY KEY(id)) DEFAULT CHARACTER SET UTF8 COLLATE `UTF8_unicode_ci` ENGINE = InnoDB');
         $this->addSql('CREATE TABLE platform_faq_role (role_id CHAR(36) NOT NULL, platformFaq_id CHAR(36) NOT NULL, INDEX IDX_D707E21E3E7645EC (platformFaq_id), INDEX IDX_D707E21ED60322AC (role_id), PRIMARY KEY(platformFaq_id, role_id)) DEFAULT CHARACTER SET UTF8 COLLATE `UTF8_unicode_ci` ENGINE = InnoDB');
-        $this->addSql('CREATE TABLE platform_faq_category (id CHAR(36) NOT NULL, title VARCHAR(255) DEFAULT \'\' NOT NULL, type VARCHAR(50) DEFAULT \'custom_category\' NOT NULL, create_date DATETIME NOT NULL, modify_date DATETIME NOT NULL, PRIMARY KEY(id)) DEFAULT CHARACTER SET UTF8 COLLATE `UTF8_unicode_ci` ENGINE = InnoDB');
+        $this->addSql('CREATE TABLE platform_faq_category (id CHAR(36) NOT NULL, title VARCHAR(255) DEFAULT \'\' NOT NULL, create_date DATETIME NOT NULL, modify_date DATETIME NOT NULL, PRIMARY KEY(id)) DEFAULT CHARACTER SET UTF8 COLLATE `UTF8_unicode_ci` ENGINE = InnoDB');
         $this->addSql('ALTER TABLE platform_faq ADD CONSTRAINT FK_5ED987DF259D9BFA FOREIGN KEY (platform_faq_category_id) REFERENCES platform_faq_category (id) ON DELETE CASCADE');
         $this->addSql('ALTER TABLE platform_faq_role ADD CONSTRAINT FK_D707E21E3E7645EC FOREIGN KEY (platformFaq_id) REFERENCES platform_faq (id) ON DELETE CASCADE');
         $this->addSql('ALTER TABLE platform_faq_role ADD CONSTRAINT FK_D707E21ED60322AC FOREIGN KEY (role_id) REFERENCES _role (_r_id) ON DELETE CASCADE');
