+++ demosplan/DemosPlanCoreBundle/Entity/PlatformFaqCategory.php
@@ -16,22 +16,6 @@ use UnexpectedValueException;
 class PlatformFaqCategory extends CoreEntity implements UuidEntityInterface, FaqCategoryInterface
 {
     /**
-     * These are allowed types, independent of the role.
-     */
-    public const FAQ_CATEGORY_TYPES_MANDATORY = [
-        'system',
-        'technische_voraussetzung',
-        'bedienung',
-        'oeb_bauleitplanung',
-        'oeb_bob',
-    ];
-
-    /**
-     * These are role-dependent types.
-     */
-    public const FAQ_CATEGORY_TYPES_OPTIONAL = 'custom_category';
-
-    /**
      * @var string|null
      *
      * @ORM\Column(type="string", length=36, options={"fixed":true})
@@ -50,15 +34,6 @@ class PlatformFaqCategory extends CoreEntity implements UuidEntityInterface, Faq
      * @ORM\Column(type="string", length=255, nullable=false, options={"default":""})
      */
     protected $title;
-
-    /**
-     * Has no function for custom categories.
-     *
-     * @var string
-     *
-     * @ORM\Column(type="string", length=50, nullable=false, options={"default":"custom_category"})
-     */
-    protected $type = self::FAQ_CATEGORY_TYPES_OPTIONAL;
 
     /**
      * @var DateTime
@@ -97,27 +72,6 @@ class PlatformFaqCategory extends CoreEntity implements UuidEntityInterface, Faq
     public function getTitle(): string
     {
         return $this->title;
-    }
-
-    public function getType(): string
-    {
-        return $this->type;
-    }
-
-    public function setType(string $type): void
-    {
-        if (!in_array($type, self::FAQ_CATEGORY_TYPES_MANDATORY, true)
-            && (self::FAQ_CATEGORY_TYPES_OPTIONAL !== $type)
-        ) {
-            throw new UnexpectedValueException(sprintf('FAQ category type has the value %s, please register this value in the entity.', $type));
-        }
-
-        $this->type = $type;
-    }
-
-    public function isCustom(): bool
-    {
-        return self::FAQ_CATEGORY_TYPES_OPTIONAL === $this->type;
     }
 
     /**
